Reset formatting on main page. Resolving merge conflicts
