+++ composer.json
@@ -25,9 +25,11 @@
             "includes/db/queries/grading.php",
             "includes/db/queries/locationsRooms.php",
             "includes/db/queries/reports.php",
+            "includes/fnc/exams.php",
             "includes/operations/include.php",
             "includes/sso.inc.php",
             "includes/templates.inc.php",
+            "includes/test.php",
             "includes/utils.inc.php"
         ]
     }
