+++ includes/operations/ExamGrades.class.php
@@ -12,21 +12,15 @@ class ExamGrades extends Operation
 {
     function __construct()
     {
-        parent::setAllowedAccountTypes(array(ACCOUNT_TYPE_ADMIN, ACCOUNT_TYPE_TEACHER));
+        parent::requireLogin(false);
 
-        parent::registerExecution(array($this, 'getExamGrades'));
+        parent::registerExecution(array($this, 'getExamInformation'));
 
-        parent::registerParameter('examID', 'integer');
+        parent::registerParameter('id', 'integer');
         parent::registerStaticParameter('archivedState', EXAM_STATE_ARCHIVED);
 
-        parent::registerAccountIDValidation(
-            'validateUserCanEditExam', 'examID'
-        );
-
-        parent::registerValidation('validateExamIDExists', 'examID');
-        parent::registerValidation(
-            'validateExamStateIs', array('examID', 'archivedState')
-        );
+        parent::registerValidation('validateExamIDExists', 'id');
+        parent::registerValidation('validateExamStateIs',array('id', 'archivedState'));
     }
 
     public function execute(array $args, string $accountID = null)
@@ -34,29 +28,18 @@ class ExamGrades extends Operation
         return parent::execute($args, $accountID);
     }
 
-    /**
-     * @param int $examID
-     *
-     * @return array        Associative array, keys
-     *                      'grades' => set of grade info
-     *                      grade info
-     *                      'studentID'
-     *                      'grade'
-     *                      'passed'
-     */
     public static function getExamGrades(int $examID)
     {
         $registrations = getExamRegistrations($examID);
         $grades = array();
-        foreach ($registrations as $studentID) {
-            $info = getStudentExamGradeDetails($examID, $studentID);
-            $grade = array('studentID' => $studentID,
-                           'grade'     => $info['grade'],
-                           'passed'    => $info['passed']);
+        foreach($registrations as $studentID) {
+            $points = getStudentExamGrade($examID, $studentID);
+            $grade = array('studentID' => $studentID, 'grade' => $points);
             array_push($grades, $grade);
         }
 
         return array(
+            'examID' => $examID,
             'grades' => $grades
         );
     }
