--- includes/operations/FinalizeExam.class.php
@@ -1,47 +1,0 @@
-<?php
-
-/**
- * To finalize an exam
- *
- * @author         Mathew McCain
- * @category       APE
- * @package        APE_includes
- * @subpackage     Operation
- */
-class FinalizeExam extends Operation
-{
-    function __construct()
-    {
-        parent::setAllowedAccountTypes(
-            array(ACCOUNT_TYPE_TEACHER, ACCOUNT_TYPE_ADMIN)
-        );
-
-        parent::registerParameter('examID', 'integer');
-
-        parent::registerStaticParameter('finalizeState', EXAM_STATE_FINALIZING);
-
-        parent::registerExecution(array($this, 'finalize'));
-
-        parent::registerAccountIDValidation(
-            'validateUserCanEditExam', 'examID'
-        );
-
-        parent::registerValidation('validateExamIDExists', 'examID');
-        parent::registerValidation(
-            'validateExamStateIs', array('examID', 'finalizeState')
-        );
-        parent::registerValidation('validateNoConflictsExist', 'examID');
-    }
-
-    public function execute(array $args, string $accountID = null)
-    {
-        return parent::execute($args, $accountID);
-    }
-
-    public static function finalize(int $examID)
-    {
-        finalizeExam($examID);
-
-        return array(true);
-    }
-}
