+++ scripts/refresh.php
@@ -1,10 +1,10 @@
-#!/usr/bin/php7.0
 <?php
 require_once '../config.php';
 
 /*
  * Script to run any necessary refresh functions
- * See 'refresh.sh' for more details
+ * set cron job
+ * 	5 * * * *
  */
  
 refreshExams();
