--- test/operation_examsearch.php
@@ -1,41 +1,0 @@
-<?php
-/**
- * file for testing queries
- *
- * @author         Mathew McCain
- * @category       APE
- * @package        APE_test
- * @subpackage     Blank
- */
-
-require_once '../config.php';
-
-echo "testing operations...</br></br>";
-
-function testOperation($args, $accountID) {
-    echo "</br>testing some operation:</br>";
-    echo "account ID: $accountID</br>";
-    echo "args: ";
-    print_r($args);
-    echo "</br>";
-
-    try {
-        $op = new ExamSearch();
-
-        $result = $op->execute($args, $accountID);
-
-        echo "results: </br>";
-        print_r($result);
-        echo "</br>";
-    } catch (Exception $e) {
-        echo "exception - $e</br>";
-    }
-    echo "</br>";
-}
-
-// 'states', 'types', 'teacherID'
-
-testOperation(array('states' => GET_EXAMS_ALL, 'types' => GET_EXAMS_TYPE_BOTH), '00111133');
-
-
-echo "</br></br>completed test</br>";
